2019.04.16 15:40

+++ smartGym-backEnd/smartGym-backEnd-dao/src/main/java/cn/smartGym/mapper/PictureMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface PictureMapper {
-    int countByExample(PictureExample example);
+    long countByExample(PictureExample example);
 
     int deleteByExample(PictureExample example);
